SkillQuest_GenAI/FINAL_Optune/requirements.txt
transformers>=4.11.0
neptune-client>=0.14.2
optuna>=2.10.0
bert-score>=0.3.11
matplotlib>=3.4.3
seaborn>=0.11.2
pandas>=1.3.3
python-dotenv>=0.20.0
